Renamed urlFromBinary.js to urlFromBinaryObj.js
